python/pip_requirements.txt
google-api-python-client==1.7.2
google-auth==1.8.0
google-auth-httplib2==0.0.3
hosker_utils
oauth2client
mega_utils
progressbar2